+++ CHANGELOG.md
@@ -14,7 +14,8 @@ README.md to use the newest tag with new release
 
 - Step `cleanup_instance_files` to clean up data of stopped instance;
 - Add availability to set environment variables for instance service;
-- Add `instances_from_same_machine` variable in preparation.
+- Add `instances_from_same_machine` variable in preparation;
+- Add `edit_topology_check` step to compare inventory and real cluster topology.
 
 ### Changed